requirements.txt
ollama
python-telegram-bot==20.0
requests
telebot
googlesearch-python
python-dotenv